SortoBotPy/PyTest/Alt/requirements.txt
pyserial
pandas


